requirements.txt
quart
python-dotenv
pydantic
tortoise-orm